+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Reformat added on date for files
 * **New "create collection" with multi-step wizard interface**
 * Ask for confirmation before deleting a collection in Collections view
+* Pause/resume video on click in player area
 
 ### Lamia
 * New service lamia, extracts media tags from files (codec, resolution, duration) and adds them to Lilin